Show JSON data in Paint View
When you load the paint catalog, the saved paints in the JSON storage file (temporarily JSON_Test.json) will be listed by name.
Plan is to have more information like rgb value and brand info etc.
Also looking to implement a detailed view in the future, which shows metadata for an item.