Show more attributes form a paint item
So now attributes like type and brand are shown together with the paint name.
Also decided to add the labels in one frame to make it easier to design the GUI in the future.
Might wanna add a bigger area on the right of Paint_View to show the expanded details there, like a color example and other meta data.