Insert JSON data into paints table.
So how this works is that the code iterates through each entry for 'Paints', which is 'v' in my code for value.
For each iteration relevant attribute values are inserted into the table through the use of 'v' variable.
This way the table can easily expand both horisontally and vertically.