Implement detailed view for paints
Selected paints will now show on the frame to the right of the paint_view window.
Detailed data such as ID and RGB values show here (ID is TBD)
A canvas with a small rectangle in it displays the color value by changing its bacground to the value listed in a paint item.